+++ src/super/i-block/modules/state/index.ts
@@ -14,7 +14,7 @@
 import symbolGenerator from 'core/symbol';
 
 import iBlock from 'super/i-block/i-block';
-import Friend from 'super/i-block/friend';
+import Friend from 'friends/friend';
 
 export * from 'super/i-block/modules/state/interface';